Fix issue #14
Added `\ekmStartTrillSpanScript` to specify any text (EXTEXT) instead of the tr symbol like an ornament with `\ekmScript`, e.g. making use of a ligature with

    \ekmStartTrillSpanScript #-1 #'(#xE262 #xE566)

Else, for a proper vertical positioning with tr, the standard accidentals for chord symbols U+ED60 ... should be used, e.g.

    \ekmStartTrillSpanScript
      #-1
      #(markup #:concat
        (#:ekm-char #xE566
         #:hspace 0.2
         #:fontsize -3 #:ekm-char #xED62
         #:hspace 0.2))

For "real" accidental glyphs, a general positioning is tricky.