Fix issue #5
This is a workaround since there is no specific scoop grob in LilyPond. The music function `\ekmScoop` is a variant of `\ekmPlayWith` and draws scoop/plop as a Fingering grob with the definition keys `S` and `P` and the aux. markup command `\ekm-scoop`. It doesn't need the SMuFL switch `fingering`.